Added seperate route to update user avatar image in serverside in profile.js Created seperate folder "common" under components and moved TextFieldGroup.js inside that folder Added Upload avatar Action, Added profile Reducer for uploadavatar, Created seperate file for uploadavatar.js in edit profile . Added Common textinput fields to Createprofile.js, Css changes in login page, landing page and createProfile.js. Uploaded image in CreatePost . Toast successful msg and navigated to profile page. Deleted CreateProfile.js Fixed CSS issues in EditProfile. DIsplaying Change avatar and password in small screens.
