+++ client/src/App.js
@@ -11,7 +11,6 @@ import Login from './components/Auth/LogIn/Login';
 import Homepage from './components/Home/Homepage';
 import Suggestion from './components/Suggestion/Suggestion';
 import EditProfile from './components/EditProfile/EditProfile';
-import CreateProfile from './components/EditProfile/CreateProfile.js';
 import ChangePassword from './components/EditProfile/ChangePassword';
 import ProfileHeader from './components/Profile/ProfileHeader';
 import CreatePost from './components/Post/CreatePost';
@@ -64,7 +63,6 @@ class App extends Component {
               <Route exact path="/home" component={Homepage} />
               <Route exact path="/suggestion" component={Suggestion} />
               <Route exact path="/edit" component={EditProfile} />
-              <Route exact path="/createprofile" component={CreateProfile} />
               <Route exact path='/changepassword' component={ChangePassword} />
               <Route exact path="/profile" component={ProfileHeader} />
               <Route exact path="/single-post" component={SinglePost} />
