+++ client/src/actions/profileActions.js
@@ -8,24 +8,6 @@ import {
 } from './types'
 
 
-//Upload Avatar
-export const uploadAvatar = (newAvatar) => dispatch => {
-
-  axios.put('/api/profile/editavatar', newAvatar)
-    .then(res => {
-      dispatch({
-        type: UPLOAD_AVATAR,
-        payload: res.data
-      })
-    })
-    .catch(err => {
-      dispatch({
-        type: GET_ERRORS,
-        payload: err.response.data
-      })
-    });
-}
-
 // Get current profile
 export const getCurrentProfile = () => dispatch => {
   dispatch(setProfileLoading());
@@ -45,6 +27,23 @@ export const getCurrentProfile = () => dispatch => {
     );
 };
 
+//Upload Avatar
+export const uploadAvatar = (newAvatar) => dispatch => {
+
+  axios.put('/api/profile/editavatar', newAvatar)
+    .then(res => {
+      dispatch({
+        type: UPLOAD_AVATAR,
+        payload: res.data
+      })
+    })
+    .catch(err => {
+      dispatch({
+        type: GET_ERRORS,
+        payload: err.response.data
+      })
+    });
+}
 
 // Delete account & profile
 export const deleteAccount = () => dispatch => {
